Intermediate single-threaded version
This change makes proof generation a blocking operation and drops the
use of separate proof generation thread.

The interface for proof generator is currently abstracted via trait that
is used in solana client logic via generics, but this is ought to change
next due to unnecessary boilerplate needed for such construct.