Process Aleo transactions in a loop
This change will add support for continuous polling of latest block and
in occurrance of a new block, it will validate its transactions and
verify the transition's proof inside each transaction.

There's also rough timing of validation and proof verification via use
of RDTSC, but this must be treated with a grain of salt. It's mostly a
rough estimation of instructions that take place during the operation,
but due to way how multi-core processors work, it's not fully reliable
in all cases.